feat(execution,audit): ParamsPolicy + integer-cents cost_delta (ACR-300@2.1)
Two spec-alignment items from TODO.md, cut together because the Tool
interface changed in both.

ParamsPolicy (Spec v0.2 Part 6)
- New execution.ParamsPolicy struct: PreviewFields (whitelist),
  SensitiveFields (mask to "*** (length: N)", rune-aware),
  HashPreviewFields (8-rune truncation), StoreHashOnly (canonical sha256).
- execution.PolicyAwareTool is an optional interface; legacy tools fall
  back to DefaultParamsPolicy preserving prior behaviour.
- maskSensitive() removed; rejection path and success path now share the
  same policy-driven masking.
- Per-tool policies declared: propose_passage whitelists bookkeeping +
  hashes content_hash; approve_draft whitelists log/draft ids + metrics;
  reject_draft/approve_agent/reject_agent/confirm_settlement_output
  whitelist their bookkeeping ids; generate_settlement +
  configure_token_rules use Default (rules tree is itself audit-worthy).

cost_delta → INTEGER cents (ACR-300 v0.2)
- Schema: audit_logs.cost_delta, covenants.budget_limit,
  budget_counters.budget_limit/spent, budget_reservations.amount all
  flipped REAL → INTEGER.
- Types: audit.Entry.CostDelta, execution.SideEffects.CostDelta,
  execution.Receipt.CostDelta, covenant.Covenant.BudgetLimit,
  budget.State.{BudgetLimit,BudgetSpent}, Tool.EstimateCost → int64.
  NetDelta stays float64 (cost_weight × cost_delta can be fractional).
- Hash chain: computeHash now branches on spec_version. Rows stamped
  ACR-300@2.0 keep %.8f format for backward-compatible verification;
  new entries stamp ACR-300@2.1 and format cost as %d. Schema default
  bumped to 2.1.
- API: /budget request body and /audit response entries carry cost as
  int64 cents.

Tests
- internal/execution/policy_test.go: 9 cases covering whitelist,
  sensitive mask, rune length, hash preview, stacked policies,
  StoreHashOnly determinism, input immutability, default legacy fields.
- internal/audit/audit_test.go: version-branched hash format (2.0 vs
  2.1 produce divergent hashes) + integer-format sensitivity guard.
- Existing suite green end-to-end.

README / TODO updated to reflect new spec_version, integer cents, and
policy-driven masking.

+++ internal/api/api.go
@@ -524,7 +524,7 @@ func (s *Server) handleGetAuditLog(w http.ResponseWriter, r *http.Request) {
 		ToolName    string  `json:"tool_name"`
 		Result      string  `json:"result"`
 		TokensDelta int     `json:"tokens_delta"`
-		CostDelta   float64 `json:"cost_delta"`
+		CostDelta   int64   `json:"cost_delta"` // USD cents
 		NetDelta    float64 `json:"net_delta"`
 		StateBefore string  `json:"state_before"`
 		StateAfter  string  `json:"state_after"`
@@ -567,7 +567,7 @@ func (s *Server) handleSetBudget(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	var req struct {
-		BudgetLimit float64 `json:"budget_limit"`
+		BudgetLimit int64 `json:"budget_limit"` // USD cents
 	}
 	if !decode(w, r, &req) {
 		return
